+++ app_tab/summary.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+def display():
+     st.write("Summary")
